Parser Error Framework
* Separate Parser
* Utilise New Parser Errors and Diagnostics
* Add additional Parser Errors
