+++ Sources/Parser/ParserError.swift
@@ -9,11 +9,19 @@ import Diagnostic
 import Source
 
 extension Diagnostic {
+  static func dummy() -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: .DUMMY, message: "Internal Error has occured")
+  }
+
   static func leftBraceExpected(in node: String, at sourceLocation: SourceLocation) -> Diagnostic {
     return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected '{' for \(node)")
   }
   static func rightBraceExpected(in node: String, at sourceLocation: SourceLocation) -> Diagnostic {
     return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected '}' for \(node)")
+  }
+
+  static func unexpectedEOF() -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: .DUMMY, message: "Unexpected end of file")
   }
 
   // MARK: Attributes
@@ -38,9 +46,43 @@ extension Diagnostic {
   static func missingContractName(at sourceLocation: SourceLocation) -> Diagnostic {
     return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected a contract name")
   }
+  static func expectedBehaviourSeparator(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected behaviour separator")
+  }
   // MARK: Struct
   static func missingStructName(at sourceLocation: SourceLocation) -> Diagnostic {
     return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected a struct name")
+  }
+
+  // MARK: Statement
+  static func expectedStatement(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected statement")
+  }
+  static func expectedForInStatement(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected 'in' between variable declaration and iterable")
+  }
+  static func statementSameLine(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Statements must be separated by a new line ")
+  }
+
+  // MARK: Components
+  static func expectedIdentifier(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected identifier")
+  }
+  static func expectedAttribute(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected '@' declared attribute")
+  }
+  static func expectedLiteral(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected literal")
+  }
+  static func expectedLeftArrow(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected '<-'")
+  }
+  static func expectedRightArrow(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected '->'")
+  }
+  static func expectedTypeAnnotation(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected type annotation")
   }
 
   // MARK: Enum
@@ -81,14 +123,26 @@ extension Diagnostic {
   static func expectedCloseParenFuncCall(at sourceLocation: SourceLocation) -> Diagnostic {
     return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected ')' to complete function-call expression")
   }
+  static func expectedCloseParen(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected ')' for bracketed expression")
+  }
   static func expectedColonAfterArgumentLabel(at sourceLocation: SourceLocation) -> Diagnostic {
     return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected ':' after argument label")
   }
   static func expectedExpr(at sourceLocation: SourceLocation) -> Diagnostic {
     return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected expression")
   }
+  static func expectedSort(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected the try to be specified with '!' or '?'")
+  }
+  static func expectedRangeOperator(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected '...' or '..<' for setting range type")
+  }
   static func expectedIdentifierAfterDot(at sourceLocation: SourceLocation) -> Diagnostic {
     return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected identifier after '.'")
+  }
+  static func expectedCloseSquareSubscript(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected ']' in subscript expression")
   }
   static func expectedCloseSquareDictionaryLiteral(at sourceLocation: SourceLocation) -> Diagnostic {
     return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected ']' in dictionary literal expression")
@@ -99,18 +153,18 @@ extension Diagnostic {
   static func expectedCloseSquareArrayLiteral(at sourceLocation: SourceLocation) -> Diagnostic {
     return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected ']' in array literal")
   }
-
+  static func expectedSeparator(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected ',' in list")
+  }
+  static func expectedEndAfterInout(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected ',' or ')' after inout identifier")
+  }
   // MARK: Generics
   static func expectedRightChevron(in node: String, at sourceLocation: SourceLocation) -> Diagnostic {
     return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected '>' to complete \(node)")
   }
   static func expectedGenericsParameterName(at sourceLocation: SourceLocation) -> Diagnostic {
     return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected an identifier to name generic parameter")
-  }
-
-  // MARK: Generics
-  static func statementSameLine(at sourceLocation: SourceLocation) -> Diagnostic {
-    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Statements must be separated by a new line ")
   }
 
   // MARK: Type
@@ -122,5 +176,8 @@ extension Diagnostic {
   }
   static func expectedCloseSquareDictionaryType(at sourceLocation: SourceLocation) -> Diagnostic {
     return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected ']' in dictionary type")
+  }
+  static func expectedIntegerInFixedArrayType(at sourceLocation: SourceLocation) -> Diagnostic {
+    return Diagnostic(severity: .error, sourceLocation: sourceLocation, message: "Expected an integer in fixed array type")
   }
 }
